# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Management API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Task Management application."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "events": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "example": "General Meeting"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Event data is ISO format and must be greater than now",
            "example": "2025-12-31T23:59:59Z"
          },
          "capacity": {
            "type": "integer",
            "minimum": 5,
            "default": 5,
            "example": 100
          },
          "registrationCount": {
            "type": "integer",
            "minimum": 0,
            "default": 0,
            "description": "cannot exceed ref:capacity",
            "maximum": "ref:capacity",
            "example": 50
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "cancelled",
              "completed"
            ],
            "default": "active"
          },
          "category": {
            "type": "string",
            "enum": [
              "conference",
              "workshop",
              "meetup",
              "seminar",
              "general"
            ],
            "default": "general"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/events": {
      "post": {
        "summary": "Create a new event",
        "tags": [
          "Events"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "example": "General Meeting"
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Event data is ISO format and must be greater than now",
                    "example": "2025-12-31T23:59:59Z"
                  },
                  "capacity": {
                    "type": "integer",
                    "minimum": 5,
                    "default": 5,
                    "example": 100
                  },
                  "registrationCount": {
                    "type": "integer",
                    "minimum": 0,
                    "default": 0,
                    "description": "cannot exceed ref:capacity",
                    "maximum": "ref:capacity",
                    "example": 50
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "cancelled",
                      "completed"
                    ],
                    "default": "active"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "conference",
                      "workshop",
                      "meetup",
                      "seminar",
                      "general"
                    ],
                    "default": "general"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/events"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/events"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve all events",
        "tags": [
          "Events"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/events"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/events"
                }
              }
            }
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "summary": "Retrieve a single event by ID",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the event",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/events"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/events"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an existing event",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the event",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/events"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/events"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}